requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain-openai==0.0.5  # For embeddings (we’ll keep this for vector store)
langchain-community==0.0.20  # For Chroma vector store
chroma==0.2.0  # For vector storage
requests==2.31.0  # For hypothetical Grok API calls